pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pool-cli"
version = "0.2.0"
description = "Simple macOS CLI to auto-categorize screenshots into Pools"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Roman Potemin" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pillow>=10.4.0",
  "numpy>=1.26.4",
  "torch>=2.3.0",
  "transformers>=4.51.3",
  "openai>=1.40.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[project.scripts]
pool = "pool_cli.main:pool_entry"

[tool.setuptools]
package-dir = { "" = "src" }
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
pool_cli = ["runtime_defaults.json"]
